refactor: inline workspace action registries
Keep workspace actions on the returned workspace object by default so create<App>Workspace exposes the registry directly. Delete one-use action factory files where the call site already owns the needed tables and runtime handles.

Update workspace docs and skills to make actions: defineActions({ ... }) the normal shape, with named registries reserved for shared action sets or registries reused before return.
