+++ apps/honeycrisp/src/routes/(signed-in)/state/folders.svelte.ts
@@ -22,7 +22,7 @@
 
 import { type FolderId, generateFolderId } from '@epicenter/honeycrisp';
 import { fromTable } from '@epicenter/svelte';
-import type { HoneycrispBrowser } from '../../../../browser';
+import type { HoneycrispBrowser } from '../../../../honeycrisp.browser';
 import { searchParams } from './search-params.svelte';
 
 export function createFolders(honeycrisp: HoneycrispBrowser) {
