+++ apps/honeycrisp/src/routes/(signed-in)/state/notes.svelte.ts
@@ -27,7 +27,7 @@ import {
 } from '@epicenter/honeycrisp';
 import { fromTable } from '@epicenter/svelte';
 import { DateTimeString } from '@epicenter/workspace';
-import type { HoneycrispBrowser } from '../../../../browser';
+import type { HoneycrispBrowser } from '../../../../honeycrisp.browser';
 import type { createFolders } from './folders.svelte';
 import { searchParams } from './search-params.svelte';
